backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
httpx<0.25.0
openai>=1.50.0

# File Processing
PyPDF2==3.0.1
python-multipart==0.0.6

# Database (for later)
sqlalchemy==2.0.23

# Development
pytest==7.4.3


chromadb==0.4.15
sentence-transformers==2.2.2
# CLIP for image embeddings
torch==2.1.1
torchvision==0.16.1
open-clip-torch==2.24.0

langgraph
langchain-core


azure-storage-blob>=12.19.0
azure-identity>=1.15.0